Crate struct_diff_patch

This crate defines traits for diffing and patching Rust structs, implements these traits for common types, and provides macros for deriving them on structs.

Modules

watch

Provides cells that can provide snapshot and update streams.

Macros

impl_simple_diff

impl_tuple_diff_patch

Structs

HashMapPatch

HashMap patches.

VecPatch

Vector patches.

Enums

Error

The common error type returned by diff and patch operations.

OptionPatch

A patch of an option.

Traits

Diff

Implements the “diff” operation, which produces a patch given two instances of the same type.

Patch

Represents a patch operating targeting values of type T.

Derive Macros

Diff

Generates a [struct_diff_patch::Diff] implementation for a struct. The patch type will match the one expected by the struct’s [struct_diff_patch::Patch] implementation, as derived by the Patch derive macro.

Patch

Derives a [struct_diff_patch::Patch] implementation for a struct. The patch type will match the one expected by the struct’s [struct_diff_patch::Diff] implementation, as derived by the Diff derive macro.