struct_diff_patch

Trait Diff

pub trait Diff: Sized {
    type Patch: Patch<Self>;

    // Required method
    fn diff(&self, other: &Self) -> Self::Patch;
}

Implements the “diff” operation, which produces a patch given two instances of the same type.

Required Associated Types

type Patch: Patch<Self>

The type of patch produced by this diff operation.

Required Methods

fn diff(&self, other: &Self) -> Self::Patch

Implements the “diff” operation, which produces a patch given two instances of the same type. Specifically, when the returned patch is applied to the original value, it should produce the second value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.