struct_diff_patch

Trait Patch

pub trait Patch<T> {
    // Required method
    fn apply(self, value: &mut T) -> Result<(), Error>;
}

Represents a patch operating targeting values of type T.

Required Methods

fn apply(self, value: &mut T) -> Result<(), Error>

Apply this patch to the provided value, consuming the patch.